requirements.txt
amqp==5.0.6
APScheduler==3.7.0
asgiref==3.3.4
billiard==3.6.4.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.4.7
defusedxml==0.7.1
Django==3.1.7
django-allauth==0.44.0
django-apscheduler==0.5.2
django-ckeditor==6.0.0
django-cleanup==5.1.0
django-js-asset==1.2.2
django-mptt==0.12.0
easy-thumbnails==2.7.1
idna==2.10
kombu==5.0.2
oauthlib==3.1.0
Pillow @ file:///Users/nikolay/PycharmProjects/py-dev-user/Pillow-8.0.1-cp38-cp38-macosx_11_0_arm64.whl
prompt-toolkit==3.0.18
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.0.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
tzlocal==2.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
